+++ examples/openapi-router/express/petstore/src/router.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import * as v from 'valibot';
-import { valibotAdapter } from 'validation-adapters/valibot';
+import { vValidator } from 'validation-adapters/valibot';
 import { createExpressOpenApiRouter } from '@ibg/openapi-router';
 
 import { type paths } from './gen/v1';
@@ -9,7 +9,7 @@ export const router: Router = Router();
 export const openApiRouter = createExpressOpenApiRouter<paths>(router);
 
 openApiRouter.get('/pet/{petId}', {
-	pathAdapter: valibotAdapter(
+	pathValidator: vValidator(
 		v.object({
 			petId: v.number()
 		})