+++ examples/openapi-router/hono/petstore/src/index.ts
@@ -1,0 +1,22 @@
+import { serve } from '@hono/node-server';
+import { Hono } from 'hono';
+
+const app = new Hono();
+
+app.get('/', (c) => {
+	return c.text('Hello Hono!');
+});
+
+app.post('/:jeff', (c) => {
+	c.req.param('jeff');
+
+	return c.json({});
+});
+
+const port = 3000;
+console.log(`Server is running on port ${port}`);
+
+serve({
+	fetch: app.fetch,
+	port
+});