+++ packages/feature-fetch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # feature-fetch
 
+## 0.0.12
+
+### Patch Changes
+
+- Renamde ValidationAdapter to Validator for more clarity
+
 ## 0.0.11
 
 ### Patch Changes